Удалила ненужные файлы и добавила файлы data.js и utils.js
