--- js/utils/util.js
@@ -1,45 +1,0 @@
-export const COMMENTS = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
-];
-
-export const NAMES = [
-  'Иван',
-  'Хуан Себастьян',
-  'Мария',
-  'Кристоф',
-  'Виктор',
-  'Юлия',
-  'Люпита',
-  'Вашингтон',
-];
-
-export const PHOTO_DESCS = [
-  'Дома',
-  'На природе',
-  'На даче',
-  'Животные',
-  'Дети',
-];
-
-export const PHOTOS_AMOUNT = 25;
-
-export const MIN_LIKES = 15;
-
-export const MAX_LIKES = 200;
-
-export const MIN_AVATAR = 1;
-
-export const MAX_AVATAR = 6;
-
-export const MIN_COMMENT = 1;
-
-export const MAX_COMMENT = 10;
-
-export const MAX_STRING_LENGTH = 140;
-
-
